Refactored reindex pipeline
